requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.37.0
pydantic-settings==2.5.0
tqdm==4.66.0
numpy==1.26.0